+++ nodeid_set.h
@@ -13,20 +13,20 @@
 struct nodeID;
 
 // The usual shuffle
-void shuffle(void *base, size_t nmemb, size_t size);
+void shuffle(void *base, int nmemb, int size);
 
-void nidset_shuffle(const struct nodeID **base, size_t nmemb);
+void nidset_shuffle(const struct nodeID **base, int nmemb);
 
-int nidset_filter(const struct nodeID **dst, size_t *dst_size, const struct nodeID **src, size_t src_size, bool(*f)(const struct nodeID *));
+int nidset_filter(const struct nodeID **dst, int *dst_size, const struct nodeID **src, int src_size, bool(*f)(const struct nodeID *));
 
 // B \ A
-int nidset_complement(const struct nodeID **dst, size_t *dst_size, const struct nodeID **bs, size_t bs_size, const struct nodeID **as, size_t as_size);
+int nidset_complement(const struct nodeID **dst, int *dst_size, const struct nodeID **bs, int bs_size, const struct nodeID **as, int as_size);
 
-bool nidset_find(size_t *i, const struct nodeID **ids, size_t ids_size, const struct nodeID *id);
+bool nidset_find(int *i, const struct nodeID **ids, int ids_size, const struct nodeID *id);
 
-int nidset_add(const struct nodeID **dst, size_t *dst_size, const struct nodeID **as, size_t as_size, const struct nodeID **bs, size_t bs_size);
+int nidset_add(const struct nodeID **dst, int *dst_size, const struct nodeID **as, int as_size, const struct nodeID **bs, int bs_size);
 
-int nidset_add_i(const struct nodeID **dst, size_t *dst_size, size_t max_size, const struct nodeID **as, size_t as_size);
+int nidset_add_i(const struct nodeID **dst, int *dst_size, int max_size, const struct nodeID **as, int as_size);
 
 #endif	/* NODEID_SET_H */
